Refactor: Move callback interfaces to a single file
The `NotifyActionCallback` and `NotifyClosedCallback` interfaces were relocated from `NotifyActionCallback.kt` to `LinuxNativeNotificationIntegration.kt` for better consolidation. This change reduces fragmentation and keeps related functionality in one place, simplifying maintenance.
