+++ knotify/src/jvmMain/kotlin/io/github/kdroidfilter/knotify/platform/linux/LinuxNativeNotificationIntegration.kt
@@ -1,5 +1,6 @@
 package io.github.kdroidfilter.knotify.platform.linux
 
+import com.sun.jna.Callback
 import com.sun.jna.Library
 import com.sun.jna.Native
 import com.sun.jna.Pointer
@@ -95,4 +96,15 @@ interface LinuxNativeNotificationIntegration : Library {
      * Stop the main loop
      */
     fun quit_main_loop()
+}
+
+@FunctionalInterface
+fun interface NotifyActionCallback : Callback {
+    fun invoke(notification: Pointer?, action: String?, user_data: Pointer?)
+}
+
+// Define the closed callback as a functional interface
+@FunctionalInterface
+fun interface NotifyClosedCallback : Callback {
+    fun invoke(notification: Pointer?, user_data: Pointer?)
 }
